+++ YRpp/SwizzleManagerClass.h
@@ -40,14 +40,14 @@ public:
 
 	virtual HRESULT __stdcall Swizzle(void** pointer) override R0;
 
-	virtual HRESULT __stdcall Fetch_Swizzle_ID(void* pointer, long* id) const override R0;
+	virtual HRESULT __stdcall Fetch_Swizzle_ID(void* pointer, LONG* id) const override R0;
 
 	virtual HRESULT __stdcall Here_I_Am(long id, void* pointer) override R0;
 
 	virtual HRESULT __stdcall Save_Interface(IStream* stream, IUnknown* pointer) override R0;
 	virtual HRESULT __stdcall Load_Interface(IStream* stream, GUID* riid, void** pointer) override R0;
 
-	virtual HRESULT __stdcall Get_Save_Size(int* psize) const override R0;
+	virtual HRESULT __stdcall Get_Save_Size(LONG* psize) const override R0;
 
 	//DTOR
 	virtual ~SwizzleManagerClass() RX;
